backend/requirements.txt
flask
flask-cors
python-dotenv
apscheduler
pandas
openpyxl
xlsxwriter
reportlab
sendgrid